refactor: the library owns tracing setup, and main asks once
The binary computed the mode of operation a second time, through the
public wrapper `is_interactive`, only so it could decide where tracing
output goes. That wrapper and `log_file_path` existed for no other
caller, so two public functions and forty lines of `main.rs` served one
decision the library was already equipped to make.

`init_tracing` moves into the library, and both helpers become private.
The boundary the crate documents is unchanged: the binary still chooses
when process-global state is installed, which matters because
`tracing_subscriber::fmt().init()` panics on a second call and a test
driving `run_cli` twice would hit it.
